requirements.txt
streamlit>=1.36.0
openai>=1.40.0
exa-py>=1.0.9
python-dotenv>=1.0.1
requests>=2.32.3
agno>=0.4.0
pydantic>=2.7.0
pandas>=2.2.2
